requirements.txt
opencv-python
matplotlib
ultralytics
numpy
Pillow
